+++ experimental/packages/opentelemetry-sdk-node/src/sdk.ts
@@ -312,6 +312,12 @@ export class NodeSDK {
             })
           );
 
+    // While SDK metrics are unstable, we require an opt-in.
+    // https://opentelemetry.io/docs/specs/semconv/otel/sdk-metrics/
+    const sdkMetricsEnabled = getBooleanFromEnv(
+      'OTEL_NODE_EXPERIMENTAL_SDK_METRICS'
+    );
+
     if (
       this._meterProviderConfig?.readers &&
       // only register if there is a reader, otherwise we waste compute/memory.
@@ -321,6 +327,7 @@ export class NodeSDK {
         resource: this._resource,
         views: this._meterProviderConfig?.views ?? [],
         readers: this._meterProviderConfig.readers,
+        sdkMetricsEnabled,
       });
 
       this._meterProvider = meterProvider;
@@ -340,11 +347,6 @@ export class NodeSDK {
 
     // Only register if there is a span processor
     if (spanProcessors.length > 0) {
-      // While SDK metrics are unstable, we require an opt-in.
-      // https://opentelemetry.io/docs/specs/semconv/otel/sdk-metrics/
-      const sdkMetricsEnabled = getBooleanFromEnv(
-        'OTEL_NODE_EXPERIMENTAL_SDK_METRICS'
-      );
       this._tracerProvider = new NodeTracerProvider({
         ...this._configuration,
         resource: this._resource,