+++ packages/sdk-metrics/src/MeterProvider.ts
@@ -26,6 +26,12 @@ export interface MeterProviderOptions {
   resource?: Resource;
   views?: ViewOptions[];
   readers?: IMetricReader[];
+
+  /**
+   * Whether to enable SDK metrics for this meter provider.
+   * @experimental This option is experimental and is subject to breaking changes in minor releases.
+   */
+  sdkMetricsEnabled?: boolean;
 }
 
 /**
@@ -50,6 +56,9 @@ export class MeterProvider implements IMeterProvider {
         const collector = new MetricCollector(this._sharedState, metricReader);
         metricReader.setMetricProducer(collector);
         this._sharedState.metricCollectors.push(collector);
+        if (options.sdkMetricsEnabled) {
+          metricReader.setMeterProvider?.(this);
+        }
       }
     }
   }